refactoring project into more files
