+++ django_glue/json.py
@@ -1,0 +1,19 @@
+from django_glue.utils import GLUE_RESPONSE_TYPES
+from django.http import JsonResponse
+
+
+def generate_json_response(status, response_type: str, message_title, message_body, additional_data=None):
+    if response_type not in GLUE_RESPONSE_TYPES:
+        raise ValueError(f'response_type "{response_type}" is not a valid, choices are {GLUE_RESPONSE_TYPES}')
+
+    return JsonResponse({
+        'type': response_type,
+        'message_title': message_title,
+        'message_body': message_body,
+        'data': additional_data
+    }, status=status)
+
+
+def generate_json_404_response():
+    return generate_json_response('404', 'error', 'Request not Found',
+                                  'The requested information, object or view you are looking for was not found.')
